feat: UI for dropbarmenu
